+++ tests/test_qc_filter.py
@@ -1,3 +1,4 @@
+import logging
 
 import pytest
 import sctoolbox.qc_filter as qc
@@ -120,11 +121,11 @@ def test_filter_cells(adata):
     assert adata.shape[0] == n_false
 
 
-def test_predict_sex(capsys, adata):
+def test_predict_sex(caplog, adata):
     # gene not in data
-    qc.predict_sex(adata, groupby='sample')
-    captured = capsys.readouterr()
-    assert "Selected gene is not present in the data. Prediction is skipped." in captured.out.strip()
+    with caplog.at_level(logging.INFO):
+        qc.predict_sex(adata, groupby='sample')
+        assert "Selected gene is not present in the data. Prediction is skipped." in caplog.records[1].message
 
     # gene in data
     qc.predict_sex(adata, gene='Xkr4', gene_column='gene', groupby='sample')